+++ Editor/ShaderReferenceAbout.cs
@@ -2,7 +2,7 @@
  * @file         ShaderReferenceAbout.cs
  * @author       Hongwei Li(taecg@qq.com)
  * @created      2018-11-10
- * @updated      2021-12-27
+ * @updated      2022-01-14
  *
  * @brief       制作名单
  */
@@ -44,8 +44,8 @@ namespace taecg.tools.shaderReference
             GUIStyle labelStyle = new GUIStyle("label");
             labelStyle.fontStyle = FontStyle.Italic;
             labelStyle.alignment = TextAnchor.MiddleCenter;
-            EditorGUILayout.LabelField("Git更新地址<https://github.com/taecg/ShaderReference.git>", labelStyle);
-            EditorGUILayout.LabelField("by taecg@qq.com  updated 2022/01/06", labelStyle);
+            EditorGUILayout.TextArea("Git更新地址<https://github.com/taecg/ShaderReference.git>", labelStyle);
+            EditorGUILayout.LabelField("by taecg@qq.com  updated 2022/01/14", labelStyle);
 
             GUIStyle style = new GUIStyle();
             style.alignment = TextAnchor.MiddleCenter;